+++ contracts/lib/SafeDecimalMath.sol
@@ -20,33 +20,33 @@
 
 pragma solidity 0.7.4;
 
-import "@openzeppelin/contracts/math/SafeMath.sol";
+import '@openzeppelin/contracts/math/SafeMath.sol';
 
 library SafeDecimalMath {
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     /* Number of decimal places in the representations. */
     uint8 public constant decimals = 18;
     uint8 public constant highPrecisionDecimals = 27;
 
     /* The number representing 1.0. */
-    uint public constant UNIT = 10**uint(decimals);
+    uint256 public constant UNIT = 10**uint256(decimals);
 
     /* The number representing 1.0 for higher fidelity numbers. */
-    uint public constant PRECISE_UNIT = 10**uint(highPrecisionDecimals);
-    uint private constant UNIT_TO_HIGH_PRECISION_CONVERSION_FACTOR = 10**uint(highPrecisionDecimals - decimals);
+    uint256 public constant PRECISE_UNIT = 10**uint256(highPrecisionDecimals);
+    uint256 private constant UNIT_TO_HIGH_PRECISION_CONVERSION_FACTOR = 10**uint256(highPrecisionDecimals - decimals);
 
     /**
      * @return Provides an interface to UNIT.
      */
-    function unit() external pure returns (uint) {
+    function unit() external pure returns (uint256) {
         return UNIT;
     }
 
     /**
      * @return Provides an interface to PRECISE_UNIT.
      */
-    function preciseUnit() external pure returns (uint) {
+    function preciseUnit() external pure returns (uint256) {
         return PRECISE_UNIT;
     }
 
@@ -59,7 +59,7 @@ library SafeDecimalMath {
      * the internal division always rounds down. This helps save on gas. Rounding
      * is more expensive on gas.
      */
-    function multiplyDecimal(uint x, uint y) internal pure returns (uint) {
+    function multiplyDecimal(uint256 x, uint256 y) internal pure returns (uint256) {
         /* Divide by UNIT to remove the extra factor introduced by the product. */
         return x.mul(y) / UNIT;
     }
@@ -77,12 +77,12 @@ library SafeDecimalMath {
      * (eg. small fractions or percentages).
      */
     function _multiplyDecimalRound(
-        uint x,
-        uint y,
-        uint precisionUnit
-    ) private pure returns (uint) {
+        uint256 x,
+        uint256 y,
+        uint256 precisionUnit
+    ) private pure returns (uint256) {
         /* Divide by UNIT to remove the extra factor introduced by the product. */
-        uint quotientTimesTen = x.mul(y) / (precisionUnit / 10);
+        uint256 quotientTimesTen = x.mul(y) / (precisionUnit / 10);
 
         if (quotientTimesTen % 10 >= 5) {
             quotientTimesTen += 10;
@@ -103,7 +103,7 @@ library SafeDecimalMath {
      * Rounding is useful when you need to retain fidelity for small decimal numbers
      * (eg. small fractions or percentages).
      */
-    function multiplyDecimalRoundPrecise(uint x, uint y) internal pure returns (uint) {
+    function multiplyDecimalRoundPrecise(uint256 x, uint256 y) internal pure returns (uint256) {
         return _multiplyDecimalRound(x, y, PRECISE_UNIT);
     }
 
@@ -119,7 +119,7 @@ library SafeDecimalMath {
      * Rounding is useful when you need to retain fidelity for small decimal numbers
      * (eg. small fractions or percentages).
      */
-    function multiplyDecimalRound(uint x, uint y) internal pure returns (uint) {
+    function multiplyDecimalRound(uint256 x, uint256 y) internal pure returns (uint256) {
         return _multiplyDecimalRound(x, y, UNIT);
     }
 
@@ -132,7 +132,7 @@ library SafeDecimalMath {
      * this is an integer division, the result is always rounded down.
      * This helps save on gas. Rounding is more expensive on gas.
      */
-    function divideDecimal(uint x, uint y) internal pure returns (uint) {
+    function divideDecimal(uint256 x, uint256 y) internal pure returns (uint256) {
         /* Reintroduce the UNIT factor that will be divided out by y. */
         return x.mul(UNIT).div(y);
     }
@@ -146,11 +146,11 @@ library SafeDecimalMath {
      * be less than 2**256. The result is rounded to the nearest increment.
      */
     function _divideDecimalRound(
-        uint x,
-        uint y,
-        uint precisionUnit
-    ) private pure returns (uint) {
-        uint resultTimesTen = x.mul(precisionUnit * 10).div(y);
+        uint256 x,
+        uint256 y,
+        uint256 precisionUnit
+    ) private pure returns (uint256) {
+        uint256 resultTimesTen = x.mul(precisionUnit * 10).div(y);
 
         if (resultTimesTen % 10 >= 5) {
             resultTimesTen += 10;
@@ -167,7 +167,7 @@ library SafeDecimalMath {
      * is evaluated, so the product of x and the standard precision unit must
      * be less than 2**256. The result is rounded to the nearest increment.
      */
-    function divideDecimalRound(uint x, uint y) internal pure returns (uint) {
+    function divideDecimalRound(uint256 x, uint256 y) internal pure returns (uint256) {
         return _divideDecimalRound(x, y, UNIT);
     }
 
@@ -179,22 +179,22 @@ library SafeDecimalMath {
      * is evaluated, so the product of x and the high precision unit must
      * be less than 2**256. The result is rounded to the nearest increment.
      */
-    function divideDecimalRoundPrecise(uint x, uint y) internal pure returns (uint) {
+    function divideDecimalRoundPrecise(uint256 x, uint256 y) internal pure returns (uint256) {
         return _divideDecimalRound(x, y, PRECISE_UNIT);
     }
 
     /**
      * @dev Convert a standard decimal representation to a high precision one.
      */
-    function decimalToPreciseDecimal(uint i) internal pure returns (uint) {
+    function decimalToPreciseDecimal(uint256 i) internal pure returns (uint256) {
         return i.mul(UNIT_TO_HIGH_PRECISION_CONVERSION_FACTOR);
     }
 
     /**
      * @dev Convert a high precision decimal to a standard decimal representation.
      */
-    function preciseDecimalToDecimal(uint i) internal pure returns (uint) {
-        uint quotientTimesTen = i / (UNIT_TO_HIGH_PRECISION_CONVERSION_FACTOR / 10);
+    function preciseDecimalToDecimal(uint256 i) internal pure returns (uint256) {
+        uint256 quotientTimesTen = i / (UNIT_TO_HIGH_PRECISION_CONVERSION_FACTOR / 10);
 
         if (quotientTimesTen % 10 >= 5) {
             quotientTimesTen += 10;
